Added 'FindResourceOfType' documentation
The main goal for FindResourceOfType is to reliably find a derived ScriptableObject type within a Resources folder, preventing direct (drag-and-drop) references. It could be argued that, with the new nested Prefabs system, one could do the same with the provider object being tied to a Prefab. Nevertheless, being code-only, I'll argument that FindResourceOfType is safer - although slower on boot times, since internally it uses Resources.LoadAll. A good / moderate usage of the convenience attribute would be, for instance, a GameData scriptableObject which holds references to all other Scriptable Assets in the project.